+++ project/AndroidManifest.xml
@@ -14,9 +14,9 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:screenOrientation="portrait"
             android:name=".SplashScreen"
-            android:theme="@android:style/Theme.NoTitleBar"     >
+            android:screenOrientation="portrait"
+            android:theme="@android:style/Theme.NoTitleBar" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -25,9 +25,12 @@
         </activity>
         <activity
             android:name=".MainActivity"
-            android:label="@string/app_name">
+            android:label="@string/app_name" >
         </activity>
-        
+        <activity
+            android:name=".LoginFragment"
+            android:label="@string/app_name" >
+        </activity>
     </application>
 
 </manifest>
