+++ project/res/layout/activity_main.xml
@@ -1,25 +1,9 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFEEEE"
-    tools:context="com.j3l11234.onetopic.MainActivity" >
+    tools:context="com.j3l11234.onetopic.TestActivity"
+    tools:ignore="MergeRootFrame" />
 
-    <android.support.v4.view.ViewPager
-        android:id="@+id/main_view_paper"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        tools:context=".MainActivity" >
-
-        <android.support.v4.view.PagerTabStrip
-            android:id="@+id/main_paper_title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="#FFCCCC"
-            android:padding="10dp"
-            android:paddingBottom="4dp"
-            android:paddingTop="4dp"
-            android:textColor="#CC3333" />
-    </android.support.v4.view.ViewPager>
-
-</RelativeLayout>
