+++ project/res/values/strings.xml
@@ -4,19 +4,17 @@
     <string name="app_name">OneTopic</string>
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
-	
     <string name="tab_title_topic">Topic</string>
     <string name="tab_title_message">Message</string>
     <string name="tab_title_me">Me</string>
-    
-    
     <string name="topic_btn_support">Support</string>
     <string name="topic_btn_against">Against</string>
     <string name="topic_btn_reply">Reply</string>
     <string name="topic_btn_favorite">Favorite</string>
-    
     <string name="profile_item_participate">My Participate</string>
     <string name="profile_item_favorite">My Favorite</string>
     <string name="profile_item_comment">My Comment</string>
     <string name="profile_item_logout">Logout</string>
+    <string name="title_activity_test">TestActivity</string>
+
 </resources>
