+++ project/src/com/j3l11234/onetopic/MainActivity.java
@@ -1,38 +1,26 @@
 package com.j3l11234.onetopic;
 
 
-
-import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
 import android.support.v7.app.ActionBarActivity;
 import android.view.Menu;
 import android.view.MenuItem;
 
+
 public class MainActivity extends ActionBarActivity {
-	private ViewPager mViewPager;
-	private SectionsPagerAdapter mAdapter;
+	private FragmentManager fragmentManager;
 	
-	private TopicFragment topicFragment;
-	private MessageFragment messageFragment;
-	private MeFragment meFragment;
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_main);
-		
-		topicFragment = new TopicFragment();
-		messageFragment = new MessageFragment();
-		meFragment = new MeFragment();
-		
-		
-		mAdapter = new SectionsPagerAdapter(getSupportFragmentManager());
-        mViewPager = (ViewPager) findViewById(R.id.main_view_paper);
-        mViewPager.setAdapter(mAdapter); 
         
+		fragmentManager = getSupportFragmentManager();
+        if (savedInstanceState == null) {
+        	fragmentManager.beginTransaction()
+					.add(R.id.container, new MainFragment()).addToBackStack(null).commit();
+		}
 	}
 
 	@Override
@@ -48,41 +36,15 @@ public class MainActivity extends ActionBarActivity {
 		// automatically handle clicks on the Home/Up button, so long
 		// as you specify a parent activity in AndroidManifest.xml.
 		int id = item.getItemId();
-		if (id == R.id.action_settings) {
+		switch (id) {
+		case R.id.action_settings:
 			return true;
+		case R.id.action_login:
+			fragmentManager.beginTransaction().replace(R.id.container, new LoginFragment()).addToBackStack(null).commit();
+			break;
+		default:
+			break;
 		}
 		return super.onOptionsItemSelected(item);
 	}
-	
-	private class SectionsPagerAdapter extends FragmentPagerAdapter {
-
-        public SectionsPagerAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        @Override
-        public android.support.v4.app.Fragment getItem(int i) {
-            switch (i) {
-                case 0: return topicFragment;
-                case 1: return messageFragment;
-                case 2: return meFragment;
-            }	
-            return null;
-        }
-
-        @Override
-        public int getCount() {
-            return 3;
-        }
-
-        @Override
-        public CharSequence getPageTitle(int position) {
-            switch (position) {
-                case 0: return getString(R.string.tab_title_topic);
-                case 1: return getString(R.string.tab_title_message);
-                case 2: return getString(R.string.tab_title_me);
-            }
-            return null;
-        }
-    }
 }
