+++ project/src/com/j3l11234/onetopic/MessageFragment.java
@@ -4,9 +4,6 @@ package com.j3l11234.onetopic;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.j3l11234.onetopic.entity.MessageItem;
-import com.j3l11234.onetopic.entity.MessageItem.MessageItemTag;
-
 import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -14,9 +11,11 @@ import android.support.v4.app.ListFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
+
+import com.j3l11234.onetopic.entity.MessageItem;
+import com.j3l11234.onetopic.entity.MessageItem.MessageItemTag;
 
 
 /**
